Use abs(int) not abs(double)
When `<math.h>` gets included `abs` is overloaded for type `double` and so `abs(toffset*2)&0x1f` attempts to do bitwise AND on a `double`, which is ill-formed. This causes a build failure with GCC 6, which now provides a conforming `<math.h>` header. The fix is to ensure `abs(int)` is used instead.